Changes to CSS HTML and img file and folder
Removed the popup, and corresponding jQuery and links from the HTML file. Removed inline styling that effected the navbar when viewing form a smaller screen size.  Added a few meta tags and a Icon image to the head element.

Added styling for the dropdown menu scrollbar and buttons that show movie times. Also made formatting changes to both CSS and HTML files to make them easier to read.